Saving Submission
Basic text adventure game - the player is trapped in a room, and must unlock the locks around them with various keys to escape. I suppose bonus features can include the fact the game is not linear (the game is not a fixed series of option - they player can make as many choices as they want until they meet the win condition), and how the game prompts the player for additional playthroughs, rather than the player needing to reset the program every time.